Rewrite World_Record
Change the code to utilize more of the Twitch API, and better sections of the Speedrun.com API.

Unfortunately, this no longer allows you to search for games that aren't being played.

Pulls information from the Twitch API to get the name of the game being played, as well as the title in order to search the speedrun API for a list of that games categories.  From there it checks the existing categories against the game name to look for a match in the title, before grabbing the first place run in that game + category combination, and outputting the game, category, time and runner name from the world record run.

Furthermore made it easier to add games to the list of accepted games (SpeedrunGame / Category Extension functions need to be updated with an if statement), and added the ability to search for category extension games.